+++ src/app/__init__.py
@@ -1,0 +1,2 @@
+from .app import init
+from .app import main
